Move `arrow-rs` interop to `arrow` module
