+++ src/arrow/array/mod.rs
@@ -1,0 +1,1 @@
+pub mod fixed_size_primitive;
