+++ src/arrow/mod.rs
@@ -1,0 +1,11 @@
+#[cfg(feature = "arrow-array")]
+pub mod array;
+
+#[cfg(feature = "arrow-buffer")]
+pub mod bitmap;
+
+#[cfg(feature = "arrow-buffer")]
+pub mod buffer;
+
+#[cfg(feature = "arrow-buffer")]
+pub mod length;
