+++ src/buffer.rs
@@ -358,48 +358,6 @@ impl<T: FixedSize> BufferMut<T> for Rc<[T]> {
     }
 }
 
-#[cfg(feature = "arrow-buffer")]
-mod arrow {
-    use super::{Buffer, BufferMut, BufferType};
-    use crate::FixedSize;
-    use arrow_buffer::{ArrowNativeType, BufferBuilder, ScalarBuffer};
-
-    /// A [BufferType] implementation for [arrow_buffer::BufferBuilder].
-    pub struct ArrowBuffer;
-
-    impl BufferType for ArrowBuffer {
-        type Buffer<T: FixedSize> = BufferBuilder<T>;
-    }
-
-    impl<T: FixedSize + ArrowNativeType> Buffer<T> for BufferBuilder<T> {
-        fn as_slice(&self) -> &[T] {
-            BufferBuilder::as_slice(self)
-        }
-    }
-
-    impl<T: FixedSize + ArrowNativeType> BufferMut<T> for BufferBuilder<T> {
-        fn as_mut_slice(&mut self) -> &mut [T] {
-            BufferBuilder::as_slice_mut(self)
-        }
-    }
-
-    /// A [BufferType] implementation for [arrow_buffer::ScalarBuffer].
-    pub struct ArrowScalarBuffer;
-
-    impl BufferType for ArrowScalarBuffer {
-        type Buffer<T: FixedSize> = ScalarBuffer<T>;
-    }
-
-    impl<T: FixedSize + ArrowNativeType> Buffer<T> for ScalarBuffer<T> {
-        fn as_slice(&self) -> &[T] {
-            self
-        }
-    }
-}
-
-#[cfg(feature = "arrow-buffer")]
-pub use arrow::*;
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -463,23 +421,5 @@ mod tests {
             <_ as Buffer<u8>>::as_bytes(&slice_array_mut),
             &[0, 2, 3, 4, 5, 6]
         );
-    }
-
-    #[test]
-    #[cfg(feature = "arrow-buffer")]
-    fn arrow() {
-        let buffer = arrow_buffer::BufferBuilder::from_iter([1, 2, 3, 4]);
-        assert_eq!(<_ as Buffer<u32>>::as_slice(&buffer), &[1, 2, 3, 4]);
-
-        let mut buffer = arrow_buffer::BufferBuilder::from_iter([1u64, 2, 3, 4]);
-        assert_eq!(
-            <_ as BufferMut<u64>>::as_mut_slice(&mut buffer),
-            &[1, 2, 3, 4]
-        );
-        <_ as BufferMut<u64>>::as_mut_slice(&mut buffer)[3] = 42;
-        assert_eq!(<_ as Buffer<u64>>::as_slice(&buffer), &[1, 2, 3, 42]);
-
-        let buffer = arrow_buffer::ScalarBuffer::from_iter([1, 2, 3, 4]);
-        assert_eq!(<_ as Buffer<u32>>::as_slice(&buffer), &[1, 2, 3, 4]);
     }
 }
